Fixed duplicated code
