+++ src/main/kotlin/com/blitzoffline/giveall/command/CommandHand.kt
@@ -1,6 +1,7 @@
 package com.blitzoffline.giveall.command
 
 import com.blitzoffline.giveall.GiveAll
+import com.blitzoffline.giveall.util.calculateBoundingBox
 import com.blitzoffline.giveall.util.msg
 import dev.triumphteam.cmd.bukkit.annotation.Permission
 import dev.triumphteam.cmd.core.BaseCommand
@@ -11,7 +12,6 @@ import dev.triumphteam.cmd.core.annotation.Suggestion
 import org.bukkit.Bukkit
 import org.bukkit.Material
 import org.bukkit.entity.Player
-import org.bukkit.util.BoundingBox
 
 @Command("giveall", alias = ["gall"])
 class CommandHand(private val plugin: GiveAll) : BaseCommand() {
@@ -33,17 +33,7 @@ class CommandHand(private val plugin: GiveAll) : BaseCommand() {
             argument == null -> players = Bukkit.getServer().onlinePlayers.toList()
             argument.toDoubleOrNull() != null -> {
                 checkRadius = true
-                val location = sender.location
-                val radius = argument.toDouble()
-                val hypotenuse = kotlin.math.sqrt(2 * radius * radius)
-                val boundingBox = BoundingBox(
-                    location.x - hypotenuse,
-                    location.y - radius,
-                    location.z - hypotenuse,
-                    location.x + hypotenuse,
-                    location.y + radius,
-                    location.z + hypotenuse
-                )
+                val boundingBox = calculateBoundingBox(sender.location, argument.toDouble())
                 players = sender.world.getNearbyEntities(boundingBox).filterIsInstance<Player>().toList()
             }
             else -> {
