+++ src/main/kotlin/com/blitzoffline/giveall/command/CommandRadius.kt
@@ -1,6 +1,7 @@
 package com.blitzoffline.giveall.command
 
 import com.blitzoffline.giveall.GiveAll
+import com.blitzoffline.giveall.util.calculateBoundingBox
 import com.blitzoffline.giveall.util.msg
 import dev.triumphteam.cmd.bukkit.annotation.Permission
 import dev.triumphteam.cmd.core.BaseCommand
@@ -11,36 +12,16 @@ import dev.triumphteam.cmd.core.annotation.Suggestion
 import org.bukkit.Material
 import org.bukkit.entity.Player
 import org.bukkit.inventory.ItemStack
-import org.bukkit.util.BoundingBox
 
 @Command("giveall", alias = ["gall"])
 class CommandRadius(private val plugin: GiveAll) : BaseCommand() {
     @SubCommand("radius")
     @Permission("giveall.use.radius")
-    fun radius(sender: Player, radius: Double?, @Suggestion("materials") material: Material?, @Optional amt: Int?) {
+    fun radius(sender: Player, radius: Double, @Suggestion("materials") material: Material, @Optional amt: Int?) {
         val settings = plugin.settings
         val messages = plugin.messages
 
-        if (material == null) {
-            messages.node("WRONG-MATERIAL").getString("&cCould not find the material you specified.").msg(sender)
-            return
-        }
-
-        if (radius == null) {
-            messages.node("WRONG-RADIUS").getString("&cRadius specified is not a number.").msg(sender)
-            return
-        }
-
-        val location = sender.location
-        val hypotenuse = kotlin.math.sqrt(2 * radius * radius)
-        val boundingBox = BoundingBox(
-            location.x-hypotenuse,
-            location.y-radius, location.z-hypotenuse,
-            location.x+hypotenuse,
-            location.y+radius,
-            location.z+hypotenuse
-        )
-
+        val boundingBox = calculateBoundingBox(sender.location, radius);
         val players= sender.world.getNearbyEntities(boundingBox)
             .filterIsInstance(Player::class.java)
             .toList()
